syncutil: make IntMap generic
Closes #103147.

This commit makes syncutil.IntMap generic, so that it can be used with
keys other than int64. This restores some of the generality lost when
the standard library's sync.Map was specialized in our fork to IntMap in
8340120c. It does so without losing the performance benefits that
specialization provided, because we use generics instead of interface
objects. This limited use of generics with no virtual function calls
performs identically to the hand-specialized version.

The use of generics also provides type safety. Callers no longer need to
cast int-like keys to and from int64. They also no longer need to cast
typed-pointers to and from unsafe.Pointer.

Benchmark performance is meaningfully unchanged:
```
name                                        old time/op    new time/op    delta
LoadMostlyHits/*syncutil.IntMap-10            2.05ns ± 6%    2.33ns ± 4%  +14.01%  (p=0.000 n=10+10)
LoadMostlyMisses/*syncutil.IntMap-10          1.82ns ± 5%    2.00ns ± 7%   +9.88%  (p=0.000 n=10+9)
LoadOrStoreBalanced/*syncutil.IntMap-10        206ns ± 1%     208ns ± 2%     ~     (p=0.059 n=8+8)
LoadOrStoreUnique/*syncutil.IntMap-10          394ns ± 3%     410ns ±10%     ~     (p=0.203 n=9+10)
LoadOrStoreCollision/*syncutil.IntMap-10      1.33ns ± 5%    1.54ns ± 4%  +16.20%  (p=0.000 n=10+10)
LoadAndDeleteBalanced/*syncutil.IntMap-10     2.41ns ± 4%    2.66ns ± 2%  +10.14%  (p=0.000 n=10+10)
LoadAndDeleteUnique/*syncutil.IntMap-10       0.65ns ±41%    0.70ns ±44%     ~     (p=0.143 n=10+10)
LoadAndDeleteCollision/*syncutil.IntMap-10    0.84ns ± 3%    0.78ns ±44%     ~     (p=0.481 n=10+10)
Range/*syncutil.IntMap-10                      982ns ± 5%    1022ns ± 1%   +4.05%  (p=0.001 n=10+10)
AdversarialAlloc/*syncutil.IntMap-10           158ns ± 0%     157ns ± 1%     ~     (p=0.722 n=7+10)
AdversarialDelete/*syncutil.IntMap-10         30.6ns ±10%    29.2ns ± 5%   -4.60%  (p=0.004 n=10+9)
DeleteCollision/*syncutil.IntMap-10           0.65ns ±56%    0.88ns ±47%  +35.02%  (p=0.007 n=10+10)

name                                        old alloc/op   new alloc/op   delta
LoadMostlyHits/*syncutil.IntMap-10             0.00B          0.00B          ~     (all equal)
LoadMostlyMisses/*syncutil.IntMap-10           0.00B          0.00B          ~     (all equal)
LoadOrStoreBalanced/*syncutil.IntMap-10        49.3B ± 8%     49.2B ± 8%     ~     (p=0.854 n=10+9)
LoadOrStoreUnique/*syncutil.IntMap-10           101B ± 5%       67B ± 6%  -33.64%  (p=0.000 n=8+9)
LoadOrStoreCollision/*syncutil.IntMap-10       0.00B          0.00B          ~     (all equal)
LoadAndDeleteBalanced/*syncutil.IntMap-10      0.00B          0.00B          ~     (all equal)
LoadAndDeleteUnique/*syncutil.IntMap-10        0.00B          0.00B          ~     (all equal)
LoadAndDeleteCollision/*syncutil.IntMap-10     0.00B          0.00B          ~     (all equal)
Range/*syncutil.IntMap-10                      0.00B          0.00B          ~     (all equal)
AdversarialAlloc/*syncutil.IntMap-10           27.0B ± 0%     27.0B ± 0%     ~     (all equal)
AdversarialDelete/*syncutil.IntMap-10          8.40B ±19%     8.00B ± 0%     ~     (p=0.211 n=10+10)
DeleteCollision/*syncutil.IntMap-10            0.00B          0.00B          ~     (all equal)

name                                        old allocs/op  new allocs/op  delta
LoadMostlyHits/*syncutil.IntMap-10              0.00           0.00          ~     (all equal)
LoadMostlyMisses/*syncutil.IntMap-10            0.00           0.00          ~     (all equal)
LoadOrStoreBalanced/*syncutil.IntMap-10         0.00           0.00          ~     (all equal)
LoadOrStoreUnique/*syncutil.IntMap-10           1.00 ± 0%      1.00 ± 0%     ~     (all equal)
LoadOrStoreCollision/*syncutil.IntMap-10        0.00           0.00          ~     (all equal)
LoadAndDeleteBalanced/*syncutil.IntMap-10       0.00           0.00          ~     (all equal)
LoadAndDeleteUnique/*syncutil.IntMap-10         0.00           0.00          ~     (all equal)
LoadAndDeleteCollision/*syncutil.IntMap-10      0.00           0.00          ~     (all equal)
Range/*syncutil.IntMap-10                       0.00           0.00          ~     (all equal)
AdversarialAlloc/*syncutil.IntMap-10            0.00           0.00          ~     (all equal)
AdversarialDelete/*syncutil.IntMap-10           0.00           0.00          ~     (all equal)
DeleteCollision/*syncutil.IntMap-10             0.00           0.00          ~     (all equal)
```

Release note: None

+++ pkg/kv/kvserver/kvadmission/kvadmission.go
@@ -358,7 +358,7 @@ func (n *controllerImpl) AdmitKVWork(
 		// If flow control is disabled or if work bypasses flow control, we still
 		// subject it above-raft, leaseholder-only IO admission control.
 		if !attemptFlowControl || !admitted {
-			storeAdmissionQ := n.storeGrantCoords.TryGetQueueForStore(int32(ba.Replica.StoreID))
+			storeAdmissionQ := n.storeGrantCoords.TryGetQueueForStore(ba.Replica.StoreID)
 			if storeAdmissionQ != nil {
 				//  NB: Even though we would know here we're bypassing admission (via
 				//  `bypassAdmission`), we still have to explicitly invoke `.Admit()`.
@@ -525,7 +525,7 @@ func (n *controllerImpl) SetTenantWeightProvider(
 				n.elasticCPUGrantCoordinator.ElasticCPUWorkQueue.SetTenantWeights(weights.Node)
 
 				for _, storeWeights := range weights.Stores {
-					q := n.storeGrantCoords.TryGetQueueForStore(int32(storeWeights.StoreID))
+					q := n.storeGrantCoords.TryGetQueueForStore(storeWeights.StoreID)
 					if q != nil {
 						if kvStoresDisabled {
 							storeWeights.Weights = nil
@@ -546,7 +546,7 @@ func (n *controllerImpl) SetTenantWeightProvider(
 func (n *controllerImpl) SnapshotIngestedOrWritten(
 	storeID roachpb.StoreID, ingestStats pebble.IngestOperationStats, writeBytes uint64,
 ) {
-	storeAdmissionQ := n.storeGrantCoords.TryGetQueueForStore(int32(storeID))
+	storeAdmissionQ := n.storeGrantCoords.TryGetQueueForStore(storeID)
 	if storeAdmissionQ == nil {
 		return
 	}
@@ -560,7 +560,7 @@ func (n *controllerImpl) FollowerStoreWriteBytes(
 	if followerWriteBytes.WriteBytes == 0 && followerWriteBytes.IngestedBytes == 0 {
 		return
 	}
-	storeAdmissionQ := n.storeGrantCoords.TryGetQueueForStore(int32(storeID))
+	storeAdmissionQ := n.storeGrantCoords.TryGetQueueForStore(storeID)
 	if storeAdmissionQ == nil {
 		return
 	}
@@ -605,7 +605,7 @@ func (n *controllerImpl) AdmitRaftEntry(
 		)
 	}
 
-	storeAdmissionQ := n.storeGrantCoords.TryGetQueueForStore(int32(storeID))
+	storeAdmissionQ := n.storeGrantCoords.TryGetQueueForStore(storeID)
 	if storeAdmissionQ == nil {
 		log.Errorf(ctx, "unable to find queue for store: %s", storeID)
 		return // nothing to do
